Split each codec into its own file
