+++ lib/avro_ex/object_container/codec/null.ex
@@ -1,0 +1,7 @@
+defmodule AvroEx.ObjectContainer.Codec.Null do
+  @behaviour AvroEx.ObjectContainer.Codec
+  @impl AvroEx.ObjectContainer.Codec
+  def encode!(data), do: data
+  @impl AvroEx.ObjectContainer.Codec
+  def decode!(data), do: data
+end
